refactor(run-adapter): relocate agent-run.ts → src/agent/run-adapter.ts
Per #632 PRD bottom-up cleanup of src/legacy/core/. The adapter that
turns AgentEvent streams into chat-style {responseText, errorMessage}
results is agent-shaped — moves alongside runtime, runners, types.

- src/legacy/core/agent-run.ts → src/agent/run-adapter.ts
- src/legacy/core/agent-run.test.ts → src/agent/run-adapter.test.ts
- 3 importers updated (app.ts, http/sessions.ts, discord/discord.ts)
- Internal import paths corrected for new location

src/legacy/core/ remaining after this PR: session-store.ts +
session-store-manager.ts, tools.ts, silent-reply.ts (#676 separate).
